+++ admin/admin_log.php
@@ -59,7 +59,7 @@ $def_forums = $all_forums;
 $def_sort = $sort_desc;
 
 // Moderators data
-if (!$mod = $datastore->get('moderators')) {
+if (!$mod = $datastore->get('moderators') and !$datastore->has('moderators')) {
     $datastore->update('moderators');
     $mod = $datastore->get('moderators');
 }
@@ -71,7 +71,7 @@ $users = array($lang['ACTS_LOG_ALL_ACTIONS'] => $all_users) + array_flip($mod['m
 unset($mod);
 
 // Forums data
-if (!$forums = $datastore->get('cat_forums')) {
+if (!$forums = $datastore->get('cat_forums') and !$datastore->has('cat_forums')) {
     $datastore->update('cat_forums');
     $forums = $datastore->get('cat_forums');
 }