+++ admin/index.php
@@ -10,19 +10,19 @@
 require __DIR__ . '/pagestart.php';
 
 // Statistics
-if (!$stats = $datastore->get('stats')) {
+if (!$stats = $datastore->get('stats') and !$datastore->has('stats')) {
     $datastore->update('stats');
     $stats = $datastore->get('stats');
 }
 
 // Files integrity check
-if (!$files_integrity_data = $datastore->get('files_integrity')) {
+if (!$files_integrity_data = $datastore->get('files_integrity') and !$datastore->has('files_integrity')) {
     $datastore->update('files_integrity');
     $files_integrity_data = $datastore->get('files_integrity');
 }
 
 // Check for updates
-if (!$update_data = $datastore->get('check_updates')) {
+if (!$update_data = $datastore->get('check_updates') and !$datastore->has('check_updates')) {
     $datastore->update('check_updates');
     $update_data = $datastore->get('check_updates');
 }